fix navbar responsive issue and seperate language component.
