+++ src/App.jsx
@@ -1,6 +1,6 @@
 import './App.css';
 /* import { Route, Routes } from 'react-router-dom'; */
-import Navbar from './components/Navbar';
+import Navbar from './components/Navbar/Navbar';
 
 function App() {
   return (
