+++ src/index.js
@@ -1,10 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
 /* import { BrowserRouter as Router } from 'react-router-dom'; */
+import 'tw-elements';
+
+import './index.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
-import 'tw-elements';
 
 ReactDOM.render(
   <React.StrictMode>
